platform-tier/mlops/training-pipeline-tasks/validation_module/requirements.txt
# pandas
# numpy
# scikit-learn
tensorflow-data-validation==1.1.0
kafka-python